Corrected many small code errors (e.g. name hiding, unused variables). Changed definition of Atom::id_ to 'int' from 'short int'. Removed 'sinlist' from base/sysfunc.cpp. Added better handling of filter index loading - several default locations are now searched before the program complains and quits.

+++ trunk/src/gui/canvas.h
@@ -88,9 +88,9 @@ class Canvas
 	// Set the internal name of the canvas
 	void setName(const char *s);
 	// Return the current height of the drawing area
-	float height();
+	double height();
 	// Return the current width of the drawing area
-	float width();
+	double width();
 	// Return whether the canvas is currently drawing
 	bool isDrawing();
 	// Return if the canvas is valid
